Inclusion of Drupal tokens to ajax calls in the Maestro template editor.

+++ maestro_task_interface.class.php
@@ -48,6 +48,10 @@ abstract class MaestroTaskInterface {
     return $this->_task_id;
   }
 
+  function setSecurityToken($token) {
+    $this->_security_token = $token;
+  }
+
   protected function _fetchTaskInformation() {
     $res = db_select('maestro_template_data', 'a');
     $res->fields('a', array('task_data'));
@@ -60,6 +64,9 @@ abstract class MaestroTaskInterface {
 
   //create task will insert the shell record of the task, and then the child class will handle the edit.
   function create() {
+    if($this->_security_token != drupal_get_token()) {
+      return array('message' => t('Illegal task creation attempt.'), 'success' => 0, 'task_id' =>0);
+    }
     $rec = new stdClass();
     $rec->template_id = $this->_template_id;
     $rec->taskname = t('New Task');
@@ -71,8 +78,8 @@ abstract class MaestroTaskInterface {
       $rec->show_in_detail = 1;
     }
     $rec->first_task = 0;
-    $rec->offset_left = $_POST['offset_left'];
-    $rec->offset_top = $_POST['offset_top'];
+    $rec->offset_left = intval($_POST['offset_left']);
+    $rec->offset_top = intval($_POST['offset_top']);
     drupal_write_record('maestro_template_data', $rec);
     $this->_task_id = $rec->id;
 
@@ -81,6 +88,9 @@ abstract class MaestroTaskInterface {
 
   //deletes the task
   function destroy() {
+    if($this->_security_token != drupal_get_token()) {
+      return array('message' => t('Illegal deletion attempt.'), 'success' => 0, 'task_id' => $this->_task_id);
+    }
     $res = db_select('maestro_queue', 'a');
     $res->fields('a', array('id'));
     $res->condition('template_data_id', $this->_task_id, '=');
@@ -134,6 +144,10 @@ abstract class MaestroTaskInterface {
 
   function edit() {
     global $base_url;
+    if($this->_security_token != drupal_get_token()) {
+      return array('message' => t('Illegal edit attempt.'), 'success' => 0, 'task_id' => $this->_task_id);
+    }
+
     $maestro_url = $base_url . '/' . drupal_get_path('module', 'maestro');
 
     $res = db_select('maestro_template_data', 'a');
@@ -169,12 +183,12 @@ abstract class MaestroTaskInterface {
       }
 
       if(module_exists('og')) {
-      $res = og_get_group_ids();
-      foreach ($res as $rec) {
-        $og = og_load($rec);
-        $og_options[$rec] = $og->label;
+        $res = og_get_group_ids();
+        foreach ($res as $rec) {
+          $og = og_load($rec);
+          $og_options[$rec] = $og->label;
+        }
       }
-    }
     }
 
     //initialize the selected array
@@ -232,6 +246,9 @@ abstract class MaestroTaskInterface {
   }
 
   function save() {
+    if($this->_security_token != drupal_get_token()) {
+      return array('message' => t('Illegal save attempt.'), 'success' => 0, 'task_id' => $this->_task_id);
+    }
     $res = db_select('maestro_template_data', 'a');
     $res->fields('a', array('id', 'task_data'));
     $res->condition('a.id', $this->_task_id, '=');
@@ -381,29 +398,32 @@ abstract class MaestroTaskInterface {
 
   //remove any next step records pertaining to this task
   function clearAdjacentLines() {
-    //RK -- had to change the logic on this delete as PDO for SQL Server was failing for some reason even though the
-    //resulting query was 100% correct.
-    $taskid=intval($this->_task_id);
-    db_query("DELETE FROM {maestro_template_data_next_step} WHERE template_data_from={$taskid} OR template_data_to={$taskid} OR template_data_to_false={$taskid}");
+    if($this->_security_token == drupal_get_token()) {
+      //RK -- had to change the logic on this delete as PDO for SQL Server was failing for some reason even though the
+      //resulting query was 100% correct.
+      $taskid=intval($this->_task_id);
+      db_query("DELETE FROM {maestro_template_data_next_step} WHERE template_data_from={$taskid} OR template_data_to={$taskid} OR template_data_to_false={$taskid}");
+    }
   }
 
   //returns an array of options for when the user right-clicks the task
   function getContextMenu() {
+    $token = drupal_get_token();
     $draw_line_msg = t('Select a task to draw the line to.');
     $options = array (
       'draw_line' => array(
         'label' => t('Draw Line'),
-        'js' => "draw_status = 1; draw_type = 1; line_start = document.getElementById('task{$this->_task_id}'); set_tool_tip('$draw_line_msg');\n"
+        'js' => "tkn = '{$token}'; draw_status = 1; draw_type = 1; line_start = document.getElementById('task{$this->_task_id}'); set_tool_tip('$draw_line_msg');\n"
       ),
       'clear_lines' => array(
         'label' => t('Clear Adjacent Lines'),
-        'js' => "clear_task_lines(document.getElementById('task{$this->_task_id}'));\n"
+        'js' => "clear_task_lines(document.getElementById('task{$this->_task_id}'), '{$token}');\n"
       ),
       'edit_task' => array(
         'label' => t('Edit Task'),
         'js' => "enable_ajax_indicator(); \$.ajax({
           type: 'POST',
-          url: ajax_url + 'MaestroTaskInterface{$this->_task_type}/{$this->_task_id}/0/edit/',
+          url: ajax_url + 'MaestroTaskInterface{$this->_task_type}/{$this->_task_id}/0/edit/{$token}',
           cache: false,
           dataType: 'json',
           success: display_task_panel,
@@ -414,7 +434,7 @@ abstract class MaestroTaskInterface {
         'label' => t('Delete Task'),
         'js' => "enable_ajax_indicator(); \$.ajax({
           type: 'POST',
-          url: ajax_url + 'MaestroTaskInterface{$this->_task_type}/{$this->_task_id}/0/destroy/',
+          url: ajax_url + 'MaestroTaskInterface{$this->_task_type}/{$this->_task_id}/0/destroy/{$token}',
           cache: false,
           success: delete_task,
           dataType: 'json',
@@ -619,21 +639,22 @@ class MaestroTaskInterfaceUnknown extends MaestroTaskInterface {
   }
 
   function getContextMenu() {
+    $token = drupal_get_token();
     $draw_line_msg = t('Select a task to draw the line to.');
     $options = array (
       'draw_line' => array(
         'label' => t('Draw Line'),
-        'js' => "draw_status = 1; draw_type = 1; line_start = document.getElementById('task{$this->_task_id}'); set_tool_tip('$draw_line_msg');\n"
+        'js' => "tkn = '{$token}'; draw_status = 1; draw_type = 1; line_start = document.getElementById('task{$this->_task_id}'); set_tool_tip('$draw_line_msg');\n"
       ),
       'clear_lines' => array(
         'label' => t('Clear Adjacent Lines'),
-        'js' => "clear_task_lines(document.getElementById('task{$this->_task_id}'));\n"
+        'js' => "clear_task_lines(document.getElementById('task{$this->_task_id}'), '{$token}');\n"
       ),
       'delete_task' => array(
         'label' => t('Delete Task'),
         'js' => "enable_ajax_indicator(); \$.ajax({
           type: 'POST',
-          url: ajax_url + 'MaestroTaskInterface{$this->_task_type}/{$this->_task_id}/0/destroy/',
+          url: ajax_url + 'MaestroTaskInterface{$this->_task_type}/{$this->_task_id}/0/destroy/{$token}',
           cache: false,
           success: delete_task,
           dataType: 'json',
@@ -676,6 +697,7 @@ class MaestroTaskInterfaceStart extends MaestroTaskInterface {
   }
 
   function getContextMenu() {
+    $token = drupal_get_token();
     $draw_line_msg = t('Select a task to draw the line to.');
     $options = array (
       'draw_line' => array(
@@ -684,7 +706,7 @@ class MaestroTaskInterfaceStart extends MaestroTaskInterface {
       ),
       'clear_lines' => array(
         'label' => t('Clear Adjacent Lines'),
-        'js' => "clear_task_lines(document.getElementById('task{$this->_task_id}'));\n"
+        'js' => "clear_task_lines(document.getElementById('task{$this->_task_id}'),'{$token}');\n"
       )
     );
 
@@ -720,10 +742,11 @@ class MaestroTaskInterfaceEnd extends MaestroTaskInterface {
   }
 
   function getContextMenu() {
+    $token = drupal_get_token();
     $options = array (
       'clear_lines' => array(
         'label' => t('Clear Adjacent Lines'),
-        'js' => "clear_task_lines(document.getElementById('task{$this->_task_id}'));\n"
+        'js' => "clear_task_lines(document.getElementById('task{$this->_task_id}'),'{$token}');\n"
       )
     );
 
@@ -769,6 +792,9 @@ class MaestroTaskInterfaceIf extends MaestroTaskInterface {
   }
 
   function save() {
+    if($this->_security_token != drupal_get_token()) {
+      return array('message' => t('Illegal save attempt.'), 'success' => 0, 'task_id' => $this->_task_id);
+    }
     $rec = new stdClass();
     $rec->id = $_POST['template_data_id'];
 
@@ -796,11 +822,12 @@ class MaestroTaskInterfaceIf extends MaestroTaskInterface {
   }
 
   function getContextMenu() {
+    $token = drupal_get_token();
     $draw_line_msg = t('Select a task to draw the line to.');
     $options = array (
       'draw_line' => array(
         'label' => t('Draw Success Line'),
-        'js' => "draw_status = 1; draw_type = 1; line_start = document.getElementById('task{$this->_task_id}'); set_tool_tip('$draw_line_msg');\n"
+        'js' => "tkn = '{$token}'; draw_status = 1; draw_type = 1; line_start = document.getElementById('task{$this->_task_id}'); set_tool_tip('$draw_line_msg');\n"
       ),
       'draw_line_false' => array(
         'label' => t('Draw Fail Line'),
@@ -808,13 +835,13 @@ class MaestroTaskInterfaceIf extends MaestroTaskInterface {
       ),
       'clear_lines' => array(
         'label' => t('Clear Adjacent Lines'),
-        'js' => "clear_task_lines(document.getElementById('task{$this->_task_id}'));\n"
+        'js' => "clear_task_lines(document.getElementById('task{$this->_task_id}'), '{$token}');\n"
       ),
       'edit_task' => array(
         'label' => t('Edit Task'),
         'js' => "enable_ajax_indicator(); \$.ajax({
           type: 'POST',
-          url: ajax_url + 'MaestroTaskInterface{$this->_task_type}/{$this->_task_id}/0/edit/',
+          url: ajax_url + 'MaestroTaskInterface{$this->_task_type}/{$this->_task_id}/0/edit/{$token}',
           cache: false,
           dataType: 'json',
           success: display_task_panel,
@@ -825,7 +852,7 @@ class MaestroTaskInterfaceIf extends MaestroTaskInterface {
         'label' => t('Delete Task'),
         'js' => "enable_ajax_indicator(); \$.ajax({
           type: 'POST',
-          url: ajax_url + 'MaestroTaskInterface{$this->_task_type}/{$this->_task_id}/0/destroy/',
+          url: ajax_url + 'MaestroTaskInterface{$this->_task_type}/{$this->_task_id}/0/destroy/{$token}',
           cache: false,
           dataType: 'json',
           success: delete_task,
@@ -859,6 +886,9 @@ class MaestroTaskInterfaceBatch extends MaestroTaskInterface {
   }
 
   function save() {
+    if($this->_security_token != drupal_get_token()) {
+      return array('message' => t('Illegal save attempt.'), 'success' => 0, 'task_id' => $this->_task_id);
+    }
     $rec = new stdClass();
     $rec->id = $_POST['template_data_id'];
     $rec->task_data = serialize(array('handler' => $_POST['handler']));
@@ -895,6 +925,9 @@ class MaestroTaskInterfaceBatchFunction extends MaestroTaskInterface {
   }
 
   function save() {
+    if($this->_security_token != drupal_get_token()) {
+      return array('message' => t('Illegal save attempt.'), 'success' => 0, 'task_id' => $this->_task_id);
+    }
     $rec = new stdClass();
     $rec->id = $_POST['template_data_id'];
     $rec->task_data = serialize(array('handler' => ($_POST['handler'] == '') ? $_POST['handler_other'] : $_POST['handler']));
@@ -930,6 +963,9 @@ class MaestroTaskInterfaceInteractiveFunction extends MaestroTaskInterface {
   }
 
   function save() {
+    if($this->_security_token != drupal_get_token()) {
+      return array('message' => t('Illegal save attempt.'), 'success' => 0, 'task_id' => $this->_task_id);
+    }
     $rec = new stdClass();
     $rec->id = $_POST['template_data_id'];
     $rec->task_data = serialize(array('handler' => ($_POST['handler'] == '') ? $_POST['handler_other'] : $_POST['handler']));
@@ -989,6 +1025,9 @@ class MaestroTaskInterfaceSetProcessVariable extends MaestroTaskInterface {
   }
 
   function save() {
+    if($this->_security_token != drupal_get_token()) {
+      return array('message' => t('Illegal save attempt.'), 'success' => 0, 'task_id' => $this->_task_id);
+    }
     $rec = new stdClass();
     $rec->id = $_POST['template_data_id'];
     $methods = $this->getSetMethods();
@@ -1059,6 +1098,9 @@ class MaestroTaskInterfaceManualWeb extends MaestroTaskInterface {
   }
 
   function save() {
+    if($this->_security_token != drupal_get_token()) {
+      return array('message' => t('Illegal save attempt.'), 'success' => 0, 'task_id' => $this->_task_id);
+    }
     $rec = new stdClass();
     $rec->id = $_POST['template_data_id'];
     $rec->task_data = serialize(array(
@@ -1093,6 +1135,9 @@ class MaestroTaskInterfaceContentType extends MaestroTaskInterface {
   }
 
   function save() {
+    if($this->_security_token != drupal_get_token()) {
+      return array('message' => t('Illegal save attempt.'), 'success' => 0, 'task_id' => $this->_task_id);
+    }
     $rec = new stdClass();
     $rec->id = $_POST['template_data_id'];
     $rec->task_data = serialize(array('content_type' => $_POST['content_type']));
@@ -1149,6 +1194,9 @@ class MaestroTaskInterfaceFireTrigger extends MaestroTaskInterface {
   }
 
   function save() {
+    if($this->_security_token != drupal_get_token()) {
+      return array('message' => t('Illegal save attempt.'), 'success' => 0, 'task_id' => $this->_task_id);
+    }
     $actions = $_POST['actions'];
     $hook = 'fire_trigger_task' . $this->_task_id;
